Fix missing context in relation hooks and findOneByRelation
- ManyToMany: pass context to create() and delete() calls in
  createAndAddIdForModelA/B and deleteAndRemoveIdFromModelA/B
- MongodbDataSource.findOneByRelation: return first(data) instead of
  the full array, matching the MemoryDataSource behavior
- Restore tsconfig rootDir to ./src (was causing stale compiled files)
- Split integration tests into separate files by concern
